+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Oneview Features supported
 - Enclosure    
 - FC network
+- Logical interconnect group
 - Scope
 - Server hardware
 - Uplink Set